add swap.mined and isolate Fill.go
